+++ waspc/cli/src/Wasp/Cli/Command/Db/Seed.hs
@@ -19,9 +19,11 @@ import Wasp.Cli.Command (Command, CommandError (CommandError), require)
 import Wasp.Cli.Command.Compile (analyze)
 import Wasp.Cli.Command.Message (cliSendMessageC)
 import Wasp.Cli.Command.Require.InWaspProject (InWaspProject (InWaspProject))
+import Wasp.Cli.Services (devEnvVars, devPorts, devUrls)
 import Wasp.Generator.DbGenerator.Operations (dbSeed)
 import qualified Wasp.Message as Msg
 import Wasp.Project.Common (generatedAppDirInWaspProjectDir)
+import Wasp.Project.PerService (server)
 
 seed :: Maybe String -> Command ()
 seed maybeUserProvidedSeedName = do
@@ -34,7 +36,9 @@ seed maybeUserProvidedSeedName = do
 
   cliSendMessageC $ Msg.Start $ "Running database seed " <> nameOfSeedToRun <> "..."
 
-  liftIO (dbSeed genProjectDir nameOfSeedToRun) >>= \case
+  let waspEnvVars = server $ devEnvVars devPorts (devUrls appSpec devPorts)
+
+  liftIO (dbSeed waspEnvVars genProjectDir nameOfSeedToRun) >>= \case
     Left errorMsg -> E.throwError $ CommandError "Database seeding failed" errorMsg
     Right () -> cliSendMessageC $ Msg.Success "Database seeded successfully!"