+++ waspc/cli/src/Wasp/Cli/Command/Test.hs
@@ -9,32 +9,36 @@ import Control.Monad.Except (throwError)
 import Control.Monad.IO.Class (liftIO)
 import StrongPath (Abs, Dir, (</>))
 import StrongPath.Types (Path')
+import qualified Wasp.AppSpec as AS
 import Wasp.Cli.Command (Command, CommandError (..), require)
 import Wasp.Cli.Command.Compile (compile)
 import Wasp.Cli.Command.Message (cliSendMessageC)
 import Wasp.Cli.Command.Require.InWaspProject (InWaspProject (InWaspProject))
 import Wasp.Cli.Command.Watch (watch)
+import Wasp.Cli.Services (devEnvVars, devPorts, devUrls)
 import qualified Wasp.Generator
 import qualified Wasp.Message as Msg
 import Wasp.Project.Common
   ( WaspProjectDir,
     generatedAppDirInWaspProjectDir,
   )
+import Wasp.Project.PerService (client)
 
 test :: [String] -> Command ()
 test [] = throwError $ CommandError "Not enough arguments" "Expected: wasp test client <args>"
-test ("client" : args) = watchAndTest $ Wasp.Generator.testWebApp args
+test ("client" : args) = watchAndTest $ \appSpec ->
+  Wasp.Generator.testWebApp (devEnvVars devPorts $ devUrls appSpec devPorts).client args
 test ("server" : _args) = throwError $ CommandError "Invalid arguments" "Server testing not yet implemented."
 test _ = throwError $ CommandError "Invalid arguments" "Expected: wasp test client <args>"
 
-watchAndTest :: (Path' Abs (Dir WaspProjectDir) -> IO (Either String ())) -> Command ()
-watchAndTest testRunner = do
+watchAndTest :: (AS.AppSpec -> Path' Abs (Dir WaspProjectDir) -> IO (Either String ())) -> Command ()
+watchAndTest makeTestRunner = do
   InWaspProject waspRoot <- require
   let outDir = waspRoot </> generatedAppDirInWaspProjectDir
 
   cliSendMessageC $ Msg.Start "Starting compilation and setup phase. Hold tight..."
 
-  (warnings, _) <- compile
+  (warnings, appSpec) <- compile
 
   cliSendMessageC $ Msg.Start "Watching for file changes and running tests ..."
 
@@ -45,7 +49,7 @@ watchAndTest testRunner = do
     -- Vitest must run from the root of the project because Vite won't resolve
     -- files outside of the project root (in this case, user src/ dir which the
     -- web app imports).
-    watchWaspProjectSource `race` testRunner waspRoot
+    watchWaspProjectSource `race` makeTestRunner appSpec waspRoot
 
   case watchOrStartResult of
     Left () -> error "This should never happen, listening for file changes should never end but it did."