+++ waspc/e2e-tests/test-outputs/snapshots/wasp-migrate-golden/wasp-app/.wasp/out/sdk/wasp/client/vite/plugins/waspConfig.ts
@@ -19,10 +19,17 @@ const forcedOptions = {
   base: "/",
   envPrefix: "REACT_APP_",
   "build.outDir": ".wasp/out/web-app/build/",
+  "server.port":
+    // This can be empty. Environment variables in the server are not set when
+    // running `wasp build`.
+    envVarAsNumber("PORT"),
+  "server.strictPort": true,
 } as const;
 
 const forcedOptionHints: Partial<Record<keyof typeof forcedOptions, string>> = {
   base: "To serve your app from a subdirectory, set `client.baseDir` in your Wasp config.",
+  "server.port":
+    "To run the client on a different port, use `wasp start --client-port <port>`.",
 };
 
 export function waspConfig(): PluginOption {
@@ -39,7 +46,8 @@ export function waspConfig(): PluginOption {
           exclude: ['wasp', '@wasp.sh/lib-auth', '@wasp.sh/lib-vite-ssr']
         },
         server: {
-          port: useUserValue(config.server?.port, 3000),
+          port: forcedOptions["server.port"],
+          strictPort: forcedOptions["server.strictPort"],
           host: useUserValue(config.server?.host, "0.0.0.0"),
         },
         envPrefix: forcedOptions["envPrefix"],
@@ -110,4 +118,16 @@ function throwIfOverridingForcedOptions(config: Record<string, any>): void {
 
 function getByPath(obj: Record<string, any>, path: string): unknown {
   return path.split(".").reduce<any>((node, segment) => node?.[segment], obj);
+}
+
+function envVarAsNumber(envName: string): number | undefined {
+  const strValue = process.env[envName];
+  if (strValue === undefined) {
+    return undefined;
+  }
+  const numValue = Number.parseInt(strValue);
+  if (Number.isNaN(numValue)) {
+    throw new Error(`Environment variable ${envName} is not a valid number.`);
+  }
+  return numValue;
 }